backend/backend/urls.py
from django.contrib import admin
from django.urls import path, include
from catalog.views import UserCreateView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path("admin/", admin.site.urls),
    path("auth/user/register/", UserCreateView.as_view(), name="user_register"),
    path("auth/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("auth/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("auth/auth/", include("rest_framework.urls")),
    path("api/", include("carrier.urls")),
]
