backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.9.1",
    "black>=26.1.0",
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.7",
    "dotenv>=0.9.9",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytz>=2025.2",
    "sqlparse>=0.5.3",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = "(catalog/).*"
exclude = '''
/(
    \.git |
    \.hg |
    \.mypy_cache |
    \.tox |
    \.venv |
    venv |
    env |
    _build |
    build |
    dist |
    media |
    swagger_ui |
    swagger |
    migrations |
    __pycache__
)/
'''
force-exclude = '''
/(
    migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
extend_skip_glob = ["*/migrations/*", "*/fixtures/*"]
known_first_party = ["catalog", "carrier", "order"]
src_paths = ["."]
filter_files = true


[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
mypy_path = "."
plugins = ["mypy_django_plugin.main"]
exclude = ["migrations/"]

[tool.django-stubs]
django_settings_module = "backend.settings"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.8",
    "isort>=7.0.0",
]
